Refactor code to use Forwardable less
Forwardable was being use in excess to its actual value, now refactored
to only be used when a class needs to delegate a responsibility to
another but the other class actually forwards that responsibility
to an object it encapsulates.

Less use of constants.

Introduced Visualizer#generate_list with proper documentation